+++ src/view/components/Intro.svelte
@@ -1,3 +1,7 @@
+<script>
+import { isPremium } from "crew-components/premium";
+
+</script>
 <div class="ui-mt-16 ui-rounded-md ui-p-2" style="background-color: hsl(var(--b3));">
    <h1 class=" ui-flex ui-flex-row ui-items-center ui-justify-center">
       <iconify-icon style="font-size: 3rem" icon="mdi:alpha" /> Alpha Suit
@@ -25,6 +29,7 @@
       </div>
 
       <div class="ui-flex ui-flex-row ui-gap-2 ui-items-center ui-justify-between">
+      {#if !isPremium()}
          <a href="https://www.patreon.com/averrin" target="_blank">
             <img
                style="border: unset;"
@@ -33,8 +38,9 @@
                src="modules/alpha-suit/assets/patreon.svg"
             />
          </a>
+      {/if}
 
-         <a class="ui-flex" href="https://discord.com/channels/772596237605011466/1018495228543516722" target="_blank">
+         <a class="ui-flex" href="https://discord.gg/Nj6UuGZGrC" target="_blank">
             <img
                class="ui-rounded-md ui-border-none"
                src="https://discordapp.com/api/guilds/772596237605011466/widget.png?style=banner2"